requirements.txt
tensorflow==2.15.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.3
pydantic==2.5.3
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.4.0
pytest==8.0.0
httpx==0.26.0
kagglehub==0.2.0